ai_content_analyst: This feature was commented for fixing code standards (#123)
* ai_content_analyst: This feature was commented for fixing code standards

* Uncomment code for AI content analysis functionality

The max-chunks was not defined with a value and caused an error when calling it since it required the value None. Additionally, there was an if statement that checked, but below it there was a condition that would execute if not. It turns out that condition did not have an else to determine whether to do it or not.

Signed-off-by: CrafteosK <105567444+CrafteosK@users.noreply.github.com>

---------

Signed-off-by: CrafteosK <105567444+CrafteosK@users.noreply.github.com>
Co-authored-by: CrafteosK <105567444+CrafteosK@users.noreply.github.com>